requirements.txt
streamlit
requests
python-dotenv
streamlit-local-storage
